Do not release prerelease components with the latest tag
Previously this did not work as expected as `semver.coerce` does not
appear to handle prerelease versions. Relates to:
https://github.com/npm/node-semver/issues/357

Instead we can use `semver.prerelease` for our purposes.

We should improve test coverage here but I suspect this action
may disappear in the near future (either with a move to a mono
repo, or by extracting the bundle size notice and using a similar
approach to services for publishing).